Gemma3-OCR-app/requirements.txt
streamlit
ollama
pillow

